feat: init poll
